remove discovery query
